refactor(graph): merge references fn into Post
